coin page

+++ src/Cryptocurrencies.Application/Common/Models/CoinHistoryModel.cs
@@ -5,7 +5,7 @@ namespace Cryptocurrencies.Application.Common.Models;
 
 public record CoinHistoryModel : IMapFrom<CoinHistoryDto>
 {
-    public decimal PriceUsd { get; init; }
+    public double PriceUsd { get; init; }
 
     public long Time { get; init; } = 0;
 }
